sora dae 2.5.1

+++ feature_wallet_api/src/main/java/jp/co/soramitsu/feature_wallet_api/domain/model/LiquidityData.kt
@@ -12,5 +12,5 @@ data class LiquidityData(
     val secondReserves: BigDecimal = BigDecimal.ZERO,
     val firstPooled: BigDecimal = BigDecimal.ZERO,
     val secondPooled: BigDecimal = BigDecimal.ZERO,
-    val sbApy: BigDecimal? = null
+    val sbApy: Double? = null
 )
